Add StopAction throw

+++ Public/Remove-InfobloxObject.ps1
@@ -25,6 +25,9 @@
         [parameter(Mandatory, ParameterSetName = 'ReferenceID')][string] $ReferenceID
     )
     if (-not $Script:InfobloxConfiguration) {
+        if ($ErrorActionPreference -eq 'Stop') {
+            throw 'You must first connect to an Infoblox server using Connect-Infoblox'
+        }
         Write-Warning -Message 'Remove-InfobloxObject - You must first connect to an Infoblox server using Connect-Infoblox'
         return
     }
